+++ config/common/errors.go
@@ -69,11 +69,12 @@ var (
 	ErrFileSchemeSupport      = errors.New("file contents source must be data URL in this spec version")
 	ErrFileAppendSupport      = errors.New("appending to files is not supported in this spec version")
 	ErrFileCompressionSupport = errors.New("file compression is not supported in this spec version")
+	ErrFileHeaderSupport      = errors.New("file HTTP headers are not supported in this spec version")
 	ErrFileSpecialModeSupport = errors.New("special mode bits are not supported in this spec version")
 	ErrGroupSupport           = errors.New("groups are not supported in this spec version")
 	ErrUserFieldSupport       = errors.New("fields other than \"name\", \"ssh_authorized_keys\", and \"password_hash\" (4.13.0+) are not supported in this spec version")
 	ErrUserNameSupport        = errors.New("users other than \"core\" are not supported in this spec version")
-	ErrKernelArgumentSupport  = errors.New("this field cannot be used for kernel arguments in this spec version; use openshift.kernel_arguments instead")
+	ErrKernelArgumentSupport  = errors.New("this section cannot be used for kernel arguments in this spec version; use openshift.kernel_arguments instead")
 
 	// Storage
 	ErrClevisSupport     = errors.New("clevis is not supported in this spec version")