+++ config/openshift/v4_14_exp/translate_test.go
@@ -475,6 +475,18 @@ func TestValidateSupport(t *testing.T) {
 									},
 									Mode: util.IntToPtr(04755),
 								},
+								{
+									Path: "/i",
+									Contents: base.Resource{
+										Source: util.StrToPtr("data:,z"),
+										HTTPHeaders: base.HTTPHeaders{
+											{
+												Name:  "foo",
+												Value: util.StrToPtr("bar"),
+											},
+										},
+									},
+								},
 							},
 							Filesystems: []base.Filesystem{
 								{
@@ -541,14 +553,15 @@ func TestValidateSupport(t *testing.T) {
 				},
 			},
 			[]entry{
+				// code
 				{report.Error, common.ErrBtrfsSupport, path.New("yaml", "storage", "filesystems", 0, "format")},
 				{report.Error, common.ErrFilesystemNoneSupport, path.New("yaml", "storage", "filesystems", 1, "format")},
-				{report.Error, common.ErrDirectorySupport, path.New("yaml", "storage", "directories", 0)},
-				{report.Error, common.ErrFileAppendSupport, path.New("yaml", "storage", "files", 1, "append")},
 				{report.Error, common.ErrFileSchemeSupport, path.New("yaml", "storage", "files", 2, "contents", "source")},
 				{report.Error, common.ErrFileSpecialModeSupport, path.New("yaml", "storage", "files", 2, "mode")},
-				{report.Error, common.ErrLinkSupport, path.New("yaml", "storage", "links", 0)},
-				{report.Error, common.ErrGroupSupport, path.New("yaml", "passwd", "groups", 0)},
+				{report.Error, common.ErrUserNameSupport, path.New("yaml", "passwd", "users", 1, "name")},
+				// filters
+				{report.Error, common.ErrKernelArgumentSupport, path.New("yaml", "kernel_arguments")},
+				{report.Error, common.ErrGroupSupport, path.New("yaml", "passwd", "groups")},
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "gecos")},
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "groups")},
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "home_dir")},
@@ -560,9 +573,10 @@ func TestValidateSupport(t *testing.T) {
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "should_exist")},
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "system")},
 				{report.Error, common.ErrUserFieldSupport, path.New("yaml", "passwd", "users", 0, "uid")},
-				{report.Error, common.ErrUserNameSupport, path.New("yaml", "passwd", "users", 1)},
-				{report.Error, common.ErrKernelArgumentSupport, path.New("yaml", "kernel_arguments", "should_exist", 0)},
-				{report.Error, common.ErrKernelArgumentSupport, path.New("yaml", "kernel_arguments", "should_not_exist", 0)},
+				{report.Error, common.ErrDirectorySupport, path.New("yaml", "storage", "directories")},
+				{report.Error, common.ErrFileAppendSupport, path.New("yaml", "storage", "files", 1, "append")},
+				{report.Error, common.ErrFileHeaderSupport, path.New("yaml", "storage", "files", 3, "contents", "http_headers")},
+				{report.Error, common.ErrLinkSupport, path.New("yaml", "storage", "links")},
 			},
 		},
 	}
@@ -574,6 +588,7 @@ func TestValidateSupport(t *testing.T) {
 				expectedReport.AddOn(entry.path, entry.err, entry.kind)
 			}
 			actual, translations, r := test.in.ToMachineConfig4_14Unvalidated(common.TranslateOptions{})
+			r.Merge(fieldFilters.Verify(actual))
 			r = confutil.TranslateReportPaths(r, translations)
 			baseutil.VerifyReport(t, test.in, r)
 			assert.Equal(t, expectedReport, r, "report mismatch")